chore: Refractor wallet setup to useWallet
